+++ src/app/api/[username]/ogp/route.ts
@@ -5,7 +5,7 @@ import {
   internalServerError,
   notFoundError
 } from "@/src/app/_client/utils/http-error";
-import { opengraphRepository } from "@/src/infrastructure/repository/opengraphRepository";
+import { opengraphRepository } from "@/src/app/_server/infrastructure/repository/opengraphRepository";
 
 export async function GET(
   _: NextRequest,
