+++ src/app/api/[username]/profile-setting/route.ts
@@ -8,7 +8,7 @@ import {
   notFoundError,
   unauthorizedError
 } from "@/src/app/_client/utils/http-error";
-import { userService } from "@/src/service/userService";
+import { userService } from "@/src/app/_server/service/userService";
 
 export async function GET(
   _: NextRequest,
