+++ src/app/api/[username]/report/reflection-time/route.ts
@@ -5,7 +5,7 @@ import {
   internalServerError,
   notFoundError
 } from "@/src/app/_client/utils/http-error";
-import { reflectionService } from "@/src/service/reflectionService";
+import { reflectionService } from "@/src/app/_server/service/reflectionService";
 
 export async function GET(
   _: NextRequest,
