+++ src/app/api/reflection/[username]/route.ts
@@ -6,7 +6,7 @@ import {
   internalServerError,
   notFoundError
 } from "@/src/app/_client/utils/http-error";
-import { reflectionService } from "@/src/service/reflectionService";
+import { reflectionService } from "@/src/app/_server/service/reflectionService";
 
 export async function GET(
   req: NextRequest,
