+++ src/app/api/reflection/pinned/[reflectionCUID]/route.ts
@@ -1,7 +1,7 @@
 import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
 import { internalServerError } from "@/src/app/_client/utils/http-error";
-import { reflectionRepository } from "@/src/infrastructure/repository/reflectionRepository";
+import { reflectionRepository } from "@/src/app/_server/infrastructure/repository/reflectionRepository";
 
 export async function PATCH(
   req: NextRequest,
