+++ src/app/api/reflection/route.ts
@@ -7,7 +7,7 @@ import {
   notFoundError,
   unauthorizedError
 } from "@/src/app/_client/utils/http-error";
-import { reflectionService } from "@/src/service/reflectionService";
+import { reflectionService } from "@/src/app/_server/service/reflectionService";
 
 export async function GET(req: NextRequest) {
   try {
