+++ src/app/api/reflection/tag-counts/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { internalServerError } from "@/src/app/_client/utils/http-error";
-import { reflectionService } from "@/src/service/reflectionService";
+import { reflectionService } from "@/src/app/_server/service/reflectionService";
 
 export async function GET() {
   try {
